add worker getter

+++ BazyDanych/BazaDanych/BazaDanych/Model/WorkerProperties.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -8,7 +9,7 @@ namespace BazaDanych.Model
 {
     internal class WorkerProperties
     {
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public int WorkerId { get; set; }
